# api.yml
openapi: 3.0.3
info:
  title: API
  description: |
    API documentation including:
    - **Auth** – login, register, refresh (public); returns JWT for protected routes.
    - **User** – profile, badges, task history, streak, resume, profile pic (JWT required).
    - **Tasks** – get assigned tasks with **user_status** (completed, viewing, rejected, not_started); submit proof (JWT required).
    - **Leaderboard** – pan-india, state, college; each with optional period (all, weekly, monthly) or dedicated /weekly and /monthly routes; entry has name, rank, xp, profile_image, id, state, college.
    - **States** – list all states (public).
    - **Colleges** – list colleges by state (public).
  version: 1.0.0
servers:
  - url: /api
    description: API base path
security:
  - BearerAuth: []
paths:
  /auth/login:
    post:
      summary: User login
      description: Authenticate with email and password. Returns JWT token and user data. Use the token in the Authorization header as Bearer &lt;token&gt; for protected routes.
      operationId: login
      tags:
        - auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                user:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  name: "John Doe"
                  email: "john@example.com"
                  state_id: "state-uuid"
                  state_name: "Maharashtra"
                  college_id: "college-uuid"
                  college_name: "Example College"
                  role: "student"
                  xp: 100
                  level: 1
        '400':
          description: Bad request - invalid input (email and password required)
          content:
            text/plain:
              schema:
                type: string
              example: "Email and password are required"
        '401':
          description: Invalid credentials
          content:
            text/plain:
              schema:
                type: string
              example: "Invalid email or password"
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
              example: "Failed to generate token"
  /auth/register:
    post:
      summary: User registration
      description: |
        Register a new user. Required: name, email, password, state_id, college_id.
        Optional: referral_code, resume (file), profile_pic (file).
        Returns JWT token and user for automatic login.
      operationId: register
      tags:
        - auth
      security: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - name
                - email
                - password
                - state_id
                - college_id
              properties:
                name:
                  type: string
                  description: User's full name
                email:
                  type: string
                  format: email
                  description: User's email (must be unique)
                password:
                  type: string
                  format: password
                  description: User's password
                state_id:
                  type: string
                  description: State ID (from GET /states)
                college_id:
                  type: string
                  description: College ID (from GET /states/{stateId}/colleges)
                referral_code:
                  type: string
                  description: Optional referral code of the referrer
                resume:
                  type: string
                  format: binary
                  description: Optional resume file (PDF recommended, max 10MB)
                profile_pic:
                  type: string
                  format: binary
                  description: Optional profile picture (JPG/PNG, max 10MB)
      responses:
        '201':
          description: User created; returns token and user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
              example:
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                user:
                  id: "660e8400-e29b-41d4-a716-446655440001"
                  name: "Jane Doe"
                  email: "jane@example.com"
                  state_id: "state-uuid"
                  college_id: "college-uuid"
                  role: "student"
                  xp: 0
                  level: 1
        '400':
          description: Bad request - missing required fields or invalid data
          content:
            text/plain:
              schema:
                type: string
              example: "Missing required fields: name, email, password, state_id, college_id are required"
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
              example: "Failed to register user"
  /auth/refresh:
    post:
      summary: Refresh token
      description: |
        Exchange an old JWT (valid or expired) for a new one. Validates the signature of the old token;
        if valid, returns a new token and user. Token can be sent in the request body as {"token": "..."}
        or in the Authorization header as Bearer &lt;token&gt;.
      operationId: refreshToken
      tags:
        - auth
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: New token and user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                user:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  name: "John Doe"
                  email: "john@example.com"
                  state_id: "state-uuid"
                  state_name: "Maharashtra"
                  college_id: "college-uuid"
                  college_name: "Example College"
                  role: "student"
                  xp: 100
                  level: 1
        '400':
          description: Bad request - token required
          content:
            text/plain:
              schema:
                type: string
              example: "Token required (body.token or Authorization: Bearer <token>)"
        '401':
          description: Invalid or expired token
          content:
            text/plain:
              schema:
                type: string
              example: "Invalid or expired token"
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
              example: "Failed to generate token"
...
  /user/resume:
    post:
      summary: Upload resume
      description: Upload a resume file for the authenticated user. Only works if user hasn't uploaded a resume during registration.
      operationId: uploadResume
      tags:
        - user
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - resume
              properties:
                resume:
                  type: string
                  format: binary
                  description: Resume file (PDF recommended, max 10MB)
      responses:
        '200':
          description: Resume uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                name: "John Doe"
                resume_url: "https://bucket.s3.amazonaws.com/resumes/user-resume.pdf"
        '400':
          description: Bad request (user already has resume or invalid file)
          content:
            text/plain:
              schema:
                type: string
              example: "User already has a resume"
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                type: string
              example: "Unauthorized"
        '404':
          description: User not found
          content:
            text/plain:
              schema:
                type: string
              example: "User not found"
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
              example: "Failed to upload resume"
    put:
      summary: Update resume
      description: Update the resume file for the authenticated user. Replaces existing resume.
      operationId: updateResume
      tags:
        - user
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - resume
              properties:
                resume:
                  type: string
                  format: binary
                  description: Resume file (PDF recommended, max 10MB)
      responses:
        '200':
          description: Resume updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                resume_url: "https://bucket.s3.amazonaws.com/resumes/user-resume-v2.pdf"
        '400':
          description: Bad request - invalid file
          content:
            text/plain:
              schema:
                type: string
              example: "Invalid file type"
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                type: string
              example: "Unauthorized"
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
              example: "Failed to update resume"
  /user/profile-pic:
    post:
      summary: Upload profile picture
      description: Upload a profile picture for the authenticated user. Only works if user hasn't uploaded a profile picture during registration.
      operationId: uploadProfilePic
      tags:
        - user
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - profile_pic
              properties:
                profile_pic:
                  type: string
                  format: binary
                  description: Profile picture (JPG/PNG, max 10MB)
      responses:
        '200':
          description: Profile picture uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                avatar_url: "https://bucket.s3.amazonaws.com/profile/user-avatar.jpg"
        '400':
          description: Bad request (user already has profile picture or invalid file)
          content:
            text/plain:
              schema:
                type: string
              example: "User already has a profile picture"
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                type: string
              example: "Unauthorized"
        '404':
          description: User not found
          content:
            text/plain:
              schema:
                type: string
              example: "User not found"
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
              example: "Failed to upload profile picture"
    put:
      summary: Update profile picture
      description: Update the profile picture for the authenticated user. Replaces existing profile picture.
      operationId: updateProfilePic
      tags:
        - user
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - profile_pic
              properties:
                profile_pic:
                  type: string
                  format: binary
                  description: Profile picture (JPG/PNG, max 10MB)
      responses:
        '200':
          description: Profile picture updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                avatar_url: "https://bucket.s3.amazonaws.com/profile/user-avatar-new.jpg"
        '400':
          description: Bad request - invalid file
          content:
            text/plain:
              schema:
                type: string
              example: "Invalid file type"
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                type: string
              example: "Unauthorized"
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
              example: "Failed to update profile picture"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  schemas:
    RefreshTokenRequest:
      type: object
      description: Old JWT to exchange for a new one (optional if sent in Authorization header)
      properties:
        token:
          type: string
          description: The old JWT token (valid or expired)
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT token for Authorization header (Bearer &lt;token&gt;)
        user:
          $ref: '#/components/schemas/User'
    RegisterResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT token for Authorization header (Bearer &lt;token&gt;)
        user:
          $ref: '#/components/schemas/User'
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        state_id:
          type: string
        state_name:
          type: string
        college_id:
          type: string
        college_name:
          type: string
        role:
          type: string
          example: student
        xp:
          type: integer
        level:
          type: integer
        coins:
          type: integer
        bio:
          type: string
        avatar_url:
          type: string
          format: uri
        resume_url:
          type: string
          format: uri
        resume_visibility:
          type: string
        referral_code:
          type: string
        referred_by_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
...